move Events API query generation into EventView class
